requirements.txt
langchain==0.3.26
langchain-google-genai==2.1.6
langchain-community==0.3.27
langgraph==0.5.1
fastapi[standard]==0.115.14
python-dotenv==1.1.1
langchain-mcp-adapters==0.1.8